Remove PatternDirectedAcyclicGraph Merge VariablesLabelsMarking.kt and ActionsPreprocessing.kt Rename :marking to :preprocessing ! User input in console doesn't working for now !
