+++ plugin/src/main/kotlin/org/jetbrains/research/plugin/PatternsStorage.kt
@@ -5,10 +5,10 @@ import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
-import org.jetbrains.research.common.PatternDirectedAcyclicGraph
+import org.jetbrains.research.common.PatternGraph
 import org.jetbrains.research.common.gumtree.PyPsiGumTree
 import org.jetbrains.research.common.gumtree.wrappers.ActionWrapper
-import org.jetbrains.research.common.jgrapht.PatternDirectedAcyclicGraph
+import org.jetbrains.research.common.jgrapht.PatternGraph
 import org.jetbrains.research.common.jgrapht.edges.PatternSpecificMultipleEdge
 import org.jetbrains.research.common.jgrapht.getWeakSubgraphIsomorphismInspector
 import org.jetbrains.research.common.jgrapht.vertices.PatternSpecificVertex
@@ -29,7 +29,7 @@ import kotlin.collections.HashMap
  */
 object PatternsStorage {
     private val patternDescriptionById = HashMap<String, String>()
-    private val patternGraphById = HashMap<String, PatternDirectedAcyclicGraph>()
+    private val patternGraphById = HashMap<String, PatternGraph>()
     private val patternEditActionsById = HashMap<String, List<Action>>()
 
     private val logger = Logger.getInstance(this::class.java)
@@ -48,10 +48,10 @@ object PatternsStorage {
                 if (entryPath.matches("^patterns/[-_.A-Za-z0-9]+/graph.dot$".toRegex())) {
                     val patternId = entryPath.split("/")[1]
                     val dotSrcStream = this::class.java.getResourceAsStream("/$entryPath")
-                    val initialGraph = PatternDirectedAcyclicGraph(dotSrcStream)
+                    val initialGraph = PatternGraph(dotSrcStream)
                     val labelsGroups = loadLabelsGroups(patternId)
-                    val patternDirectedAcyclicGraph: PatternDirectedAcyclicGraph =
-                        PatternDirectedAcyclicGraph(initialGraph, labelsGroups)
+                    val patternDirectedAcyclicGraph: PatternGraph =
+                        PatternGraph(initialGraph, labelsGroups)
                     patternGraphById[patternId] = patternDirectedAcyclicGraph
                 }
             }
@@ -62,7 +62,7 @@ object PatternsStorage {
         }
     }
 
-    fun getPatternById(patternId: String): PatternDirectedAcyclicGraph? = patternGraphById[patternId]
+    fun getPatternById(patternId: String): PatternGraph? = patternGraphById[patternId]
 
     fun getPatternDescriptionById(patternId: String): String =
         patternDescriptionById.getOrPut(patternId) { loadDescription(patternId) ?: "Unnamed pattern: $patternId" }
@@ -70,7 +70,7 @@ object PatternsStorage {
     fun getPatternEditActionsById(patternId: String): List<Action> =
         patternEditActionsById.getOrPut(patternId) { loadEditActionsFromPattern(patternId) }
 
-    fun getIsomorphicPatterns(targetDirectedAcyclicGraph: PatternDirectedAcyclicGraph)
+    fun getIsomorphicPatterns(targetDirectedAcyclicGraph: PatternGraph)
             : HashMap<String, ArrayList<GraphMapping<PatternSpecificVertex, PatternSpecificMultipleEdge>>> {
         val suitablePatterns =
             HashMap<String, ArrayList<GraphMapping<PatternSpecificVertex, PatternSpecificMultipleEdge>>>()
